Disable also stops logging to files. Use disable at the end of the test to be able to delete the experiment directory afterwards.

+++ tests/test_flexp_override.py
@@ -12,11 +12,16 @@ from flexp import flexp
 
 
 def test_override():
+    expdir = path.join("tests/data/", "exp01")
+
+    # Remove the experiment dir if it exists
+    if os.path.exists(expdir):
+        shutil.rmtree(expdir)
+
     # We have to reset the _eh to make flexp stop complaining about calling setup twice.
     flexp.core._eh = {}
     flexp.setup("tests/data/", "exp01", False, override_dir=False)
 
-    expdir = path.join("tests/data/", "exp01")
     assert path.isdir(expdir), "flexp didn't create experiment dir with override_dir=False"
 
     # Test that it fails to create the directory, there should be logging file already.
@@ -28,6 +33,9 @@ def test_override():
     flexp.core._eh = {}
     flexp.setup("tests/data/", "exp01", False, override_dir=True)
 
-    # Remove the experiment dir
-    shutil.rmtree(expdir)
+    # Disable logging to be able to delete the experiment directory.
+    flexp.disable()
 
+    # Remove the experiment dir
+    if os.path.exists(expdir):
+        shutil.rmtree(expdir)
